rustPlatform.fetchCargoVendor: split script into two parts
